backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pandas>=2.0.0
requests>=2.30.0
python-multipart>=0.0.6
PyYAML>=6.0
python-dotenv>=1.0.0
pyodbc>=4.0.39
pydantic>=2.0.0
email-validator>=2.0.0
jinja2>=3.1.2 